allow uninstall mod template

+++ src/MuseDashModTools/ViewModels/Panels/Modding/ModDevelopPanelViewModel.cs
@@ -3,7 +3,7 @@
 public sealed partial class ModDevelopPanelViewModel : ViewModelBase
 {
     [ObservableProperty]
-    [NotifyCanExecuteChangedFor(nameof(InstallModTemplateCommand))]
+    [NotifyCanExecuteChangedFor(nameof(ToggleModTemplateInstallCommand))]
     public partial bool DotNetSdkInstalled { get; set; }
 
     [ObservableProperty]
@@ -20,11 +20,18 @@ public sealed partial class ModDevelopPanelViewModel : ViewModelBase
     [RelayCommand]
     private async Task InstallDotNetSdkAsync()
     {
+        var result = await MessageBoxService.NoticeConfirmOverlayAsync("Are you sure you want to install the DotNet SDK?").ConfigureAwait(true);
+        if (result is not MessageBoxResult.Yes)
+        {
+            return;
+        }
+
         Logger.ZLogInformation($"Installing DotNet SDK...");
-        var success = await PlatformService.InstallDotNetSdkAsync().ConfigureAwait(false);
+        var success = await PlatformService.InstallDotNetSdkAsync().ConfigureAwait(true);
         if (!success)
         {
             await MessageBoxService.ErrorAsync("Failed to install DotNet SDK").ConfigureAwait(false);
+            return;
         }
 
         Logger.ZLogInformation($"DotNet SDK installed successfully");
@@ -32,12 +39,21 @@ public sealed partial class ModDevelopPanelViewModel : ViewModelBase
     }
 
     [RelayCommand(CanExecute = nameof(DotNetSdkInstalled))]
+    private Task ToggleModTemplateInstallAsync() =>
+        !ModTemplateInstalled ? InstallModTemplateAsync() : UninstallModTemplateAsync();
+
     private async Task InstallModTemplateAsync()
     {
+        var result = await MessageBoxService.NoticeConfirmOverlayAsync("Are you sure you want to install the Mod Template?").ConfigureAwait(true);
+        if (result is not MessageBoxResult.Yes)
+        {
+            return;
+        }
+
         Logger.ZLogInformation($"Installing Mod Template...");
         try
         {
-            await PlatformService.InstallModTemplateAsync().ConfigureAwait(false);
+            await PlatformService.InstallModTemplateAsync().ConfigureAwait(true);
             Logger.ZLogInformation($"Mod Template installed successfully");
             ModTemplateInstalled = true;
         }
@@ -45,6 +61,28 @@ public sealed partial class ModDevelopPanelViewModel : ViewModelBase
         {
             Logger.ZLogError(ex, $"Failed to install Mod Template");
             await MessageBoxService.ErrorAsync("Failed to install Mod Template").ConfigureAwait(false);
+        }
+    }
+
+    private async Task UninstallModTemplateAsync()
+    {
+        var result = await MessageBoxService.NoticeConfirmOverlayAsync("Are you sure you want to uninstall the Mod Template?").ConfigureAwait(true);
+        if (result is not MessageBoxResult.Yes)
+        {
+            return;
+        }
+
+        Logger.ZLogInformation($"Uninstalling Mod Template...");
+        try
+        {
+            await PlatformService.UninstallModTemplateAsync().ConfigureAwait(true);
+            Logger.ZLogInformation($"Mod Template uninstalled successfully");
+            ModTemplateInstalled = false;
+        }
+        catch (Exception ex)
+        {
+            Logger.ZLogError(ex, $"Failed to uninstall Mod Template");
+            await MessageBoxService.ErrorAsync("Failed to uninstall Mod Template").ConfigureAwait(false);
         }
     }
